requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
alembic>=1.7.1
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
email-validator>=1.1.3
python-jose[cryptography]>=3.3.0
opencv-python>=4.5.3
numpy>=1.21.2
tensorflow>=2.6.0